+++ project-loader/src/test/kotlin/ArgParsingTest.kt
@@ -1,8 +1,8 @@
 import com.xenomachina.argparser.ArgParser
-import de.itemis.mps.gradle.logging.LogLevel
 import de.itemis.mps.gradle.project.loader.EnvironmentArgs
 import org.junit.Assert.assertEquals
 import org.junit.Test
+import java.util.logging.Level
 
 class ArgParsingTest {
 
@@ -11,7 +11,7 @@ class ArgParsingTest {
         val parsed = ArgParser(arrayOf(
             "--log-level", "info")).parseInto(::EnvironmentArgs)
 
-        assertEquals(LogLevel.INFO, parsed.logLevel)
+        assertEquals(Level.INFO, parsed.logLevel)
     }
 
 }